requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Talisman==1.1.0
WTForms==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
google-generativeai==0.3.2
PyPDF2==3.0.1
PyMuPDF==1.23.26
python-magic==0.4.27
email-validator==2.0.0
